[feat] 도메인 수정, 계정 관련 api, fcm 관련 api 및 서버 추가 설정

+++ src/main/java/com/epi/epilog/app/dto/DiabetesRequestDto.java
@@ -1,7 +1,6 @@
 package com.epi.epilog.app.dto;
 
 import com.epi.epilog.app.domain.annotations.ValidOccurenceType;
-import com.epi.epilog.app.domain.enums.OccurrenceType;
 import jakarta.validation.constraints.NotNull;
 import lombok.*;
 
@@ -20,6 +19,6 @@ public class DiabetesRequestDto {
         @ValidOccurenceType
         public String occurrenceType;
         @NotNull
-        public Integer bloodSugar;
+        public Double bloodSugar;
     }
 }
